18:00 commit

+++ src/main/java/com/chicken/project/store/model/service/StoreService.java
@@ -33,10 +33,16 @@ public class StoreService {
         int result = storeMapper.deleteBalanceStore(storeName);
 
         if(!(result > 0)){
+
             throw new StoreDeleteException("가맹점 정보 삭제에 실패하셨습니다.");
         } else {
 
-            storeMapper.deleteStore(storeName);
+            int result2 = storeMapper.deleteStoreRole(storeName);
+
+            if(result2 > 0){
+
+                storeMapper.deleteStore(storeName);
+            }
         }
     }
 
@@ -50,7 +56,12 @@ public class StoreService {
             throw new StoreInsertException("가맹점 등록 실패!");
         } else {
 
-            storeMapper.insertBalance(store.getStoreName());
+            int result2 = storeMapper.insertBalance(store.getStoreName(), store.getStoreId());
+
+            if(result2 > 0){
+
+                storeMapper.insertStoreRole(store);
+            }
         }
     }
 
@@ -66,7 +77,7 @@ public class StoreService {
 
         int result = storeMapper.updateStore(store);
 
-        if(!(result > 0)){
+        if(!(result > 0)) {
 
             throw new StoreUpdateException("가맹점 정보 수정 실패!");
         }
